feat(database): Registra a classificação completa no banco de dados
Adiciona persistência de fabricantes e TIPI (novos serviços)
Persiste resultado completo da classificação (service + task)
Ajusta schemas para modelo persistido
Atualiza Dockerfile, run_celery e requirements para suportar worker/DB

+++ app/__init__.py
@@ -10,7 +10,7 @@ from app.core.logger_config import logger
 from app.models.models import *
 
 
-def create_app(container: Container | None = None) -> Flask:
+def create_app(container: Container | None = None, celery_worker = False) -> Flask:
     app = Flask(__name__)
     logger.info("Iniciando aplicação flask")
     
@@ -33,19 +33,20 @@ def create_app(container: Container | None = None) -> Flask:
         
     db.init_app(app)
     with app.app_context():
-        db.drop_all()
-        db.create_all()
+        if not celery_worker:
+            db.drop_all()
+            db.create_all()
 
-        # Esperar um pouco o banco estabilizar
-        for i in range(5):
-            try:
-                seed_db()
-                print("✅ Banco populado com sucesso!")
-                break
-            except OperationalError as e:
-                print(f"⏳ Tentando novamente ({i+1}/5)...")
-                time.sleep(2)
-        else:
-            print("❌ Falha ao popular o banco após várias tentativas.")
+            # Esperar um pouco o banco estabilizar
+            for i in range(5):
+                try:
+                    seed_db()
+                    print("✅ Banco populado com sucesso!")
+                    break
+                except OperationalError as e:
+                    print(f"⏳ Tentando novamente ({i+1}/5)...")
+                    time.sleep(2)
+            else:
+                print("❌ Falha ao popular o banco após várias tentativas.")
 
     return app
